fenrir_vk_1.91
Bug fix, new Lineage theme

+++ app/src/main/java/dev/ragnarok/fenrir/spots/SpotsDialog.java
@@ -103,7 +103,7 @@ public class SpotsDialog extends AlertDialog {
             @SuppressLint("Recycle") Animator move = ObjectAnimator.ofFloat(animatedView, "xFactor", 0, 1);
             move.setDuration(DURATION);
             move.setInterpolator(new HesitateInterpolator());
-            move.setStartDelay(DELAY * i);
+            move.setStartDelay((long) DELAY * i);
             move.addListener(new AnimatorListenerAdapter() {
                 @Override
                 public void onAnimationEnd(Animator animation) {
